Move inner types to the next to relavant files
