+++ src/redux/domain/card.g.ts
