+++ src/redux/domain/column.g.ts
@@ -1,4 +1,4 @@
-import { Card } from "src/types/inner/card.g";
+import { Card } from "src/redux/domain/card.g";
 
 export interface Column {
   id: number;
