+++ src/redux/domain/column.ts
@@ -1,5 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { Column as innerColumn } from "src/types/inner/column.g";
+import { Column as innerColumn } from "src/redux/domain/column.g";
 import { Column as OuterColumn } from "src/api/drello-api/board";
 import { RootState } from "src/redux/root";
 import { getBoardsThunk } from "src/redux/domain/board";
