+++ src/redux/domain/workspace.g.ts
